27: WIP: renamed SquareCard to SimpleHoverCard
