+++ src/components/SimpleHoverCard/SimpleHoverCard.module.scss
