+++ src/components/SimpleHoverCard/SimpleHoverCard.module.scss.d.ts
