+++ src/components/SimpleHoverCard/SimpleHoverCard.story.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+
+import { storiesOf } from "@storybook/react";
+
+import { SimpleHoverCard } from "./index";
+import { projectBox } from "../../types/fixtures";
+
+storiesOf("SimpleHoverCard", module).add("Default", () => (
+    <SimpleHoverCard className="myClass" project={projectBox} />
+));
