+++ src/components/SimpleHoverCard/index.ts
@@ -1,0 +1,1 @@
+export { SimpleHoverCard } from "./SimpleHoverCard";
