--- src/components/SquareCard/SquareCard.story.tsx
@@ -1,14 +1,0 @@
-
-import * as React from "react";
-
-
-import { storiesOf } from "@storybook/react";
-
-import { SquareCard } from "./index";
-
-
-storiesOf("SquareCard", module).add(
-    "Default", 
-    () => (
-    <SquareCard  className="myClass" />
-));
