--- src/components/SquareCard/index.ts
@@ -1,1 +1,0 @@
-export { SquareCard } from "./SquareCard";
