+++ src/components/SquareCarousel/SquareCarousel.module.scss.d.ts
@@ -1,3 +1,4 @@
 /* tslint:disable */
 // This file is automatically generated from your CSS. Any edits will be overwritten.
 export const component: string;
+export const card: string;
